Migrate func_w0612.py to new functional tests
