+++ tests/functional/u/unused/unused_variable.txt
@@ -12,3 +12,9 @@ unused-import:54:4:unused_import_from:Unused wraps imported from functools as ab
 unused-import:55:4:unused_import_from:Unused namedtuple imported from collections
 unused-import:59:4:unused_import_in_function:Unused hexdigits imported from string
 unused-variable:64:4:hello:Unused variable 'my_var'
+unused-variable:76:4:function:Unused variable 'aaaa'
+global-statement:97:4:test_global:Using the global statement
+unused-import:103:4:test_global:Unused platform imported from sys
+unused-import:104:4:test_global:Unused version imported from sys as VERSION
+unused-import:105:4:test_global:Unused import this
+unused-import:106:4:test_global:Unused re imported as RE
